Move match_url_pattern to a helper file
